added changes to the nba constructor and cleaned the files up
